requirements.txt
chainlit
langgraph
langchain
langchain-community
langchain-openai
easyocr
pdf2image
pytesseract
sentence_transformers
python-dotenv
pymupdf